requirements.txt
asgiref==3.3.4
Django==3.2
djangorestframework==3.12.4
python-dateutil==2.8.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
